fix: address CodeRabbit review comments on PR #141
- pyproject.toml: license metadata MIT -> AGPL-3.0 to match LICENSE (AGPLv3
  since PR #137). Metadata was stale, not specific to this PR, but directly
  relates to the licensing claims in this PR's description.
- trapdata/ml/models/localization.py: drop int32 cast before cv2.minAreaRect.
  cv2 accepts float points and the cast was dropping ~0.2 deg of rotation
  precision on shallow-angled boxes. Verified empirically.
- trapdata/api/models/classification.py: widen detector_cls annotation from
  type[APIMothDetector] to type[APIInferenceBaseClass]. APIMothDetector_YOLO11m_Mothbot
  is a sibling of APIMothDetector, not a subclass — common ancestor is
  APIInferenceBaseClass.

+++ trapdata/ml/models/localization.py
@@ -395,7 +395,7 @@ def _corners_to_yolo_detection(corners: np.ndarray, score: float) -> YoloDetecti
     pts = np.asarray(corners, dtype=np.float32).reshape(-1, 2)
     x1, y1 = float(pts[:, 0].min()), float(pts[:, 1].min())
     x2, y2 = float(pts[:, 0].max()), float(pts[:, 1].max())
-    rect = cv2.minAreaRect(pts.astype(np.int32))
+    rect = cv2.minAreaRect(pts)
     angle = float(rect[2])
     return YoloDetection(x1=x1, y1=y1, x2=x2, y2=y2, rotation=angle, score=float(score))
